Socket connection to the software containing the ANN
The socket connection is working between 'Analize' and the ANN software (Brain).  Analize sends the filename and path of the picture to Brain in a raw format. Brain formats the data, and creates a dataset from the .csv file, which has the same path and name as the picture. (I have also changed the .txt format to .csv, as Python Pandas can only import data in .csv format.)
The connection is also working to RamanComm, as it is not on my notebook, I haven't tested if both are working parallel.